Rename RPC compatibility modules to avoid confusion
